+++ packages/adapter-aphrodite/tests/AphroditeAesthetic.test.ts
@@ -50,12 +50,11 @@ describe('AphroditeAesthetic', () => {
     const convertedSheet = global
       ? instance.syntax.convertGlobalSheet(styleSheet, options).toObject()
       : instance.syntax.convertStyleSheet(styleSheet, options).toObject();
-    // @ts-ignore Allow access
-    const styles = instance.processStyleSheet(convertedSheet, 'aphrodite');
+    const parsedSheet = instance.parseStyleSheet(convertedSheet);
 
     expect(convertedSheet).toEqual(convertDirection(expectedStyles, dir));
 
-    expect(instance.transformStyles(Object.values(styles), options)).toBe(expectedClassName);
+    expect(instance.transformStyles(Object.values(parsedSheet), options)).toBe(expectedClassName);
 
     if (raw) {
       // @ts-ignore