+++ packages/core/CHANGELOG.md
@@ -12,6 +12,8 @@ To support RTL mode and other frameworks besides React, the core APIs had to sli
   of being spread across all arguments.
 - Updated `Aesthetic#transformToClassName` to only be passed the parsed blocks and not the native
   adapter blocks.
+- Renamed `Aesthetic#setStyleSheet` to `registerStyleSheet`.
+- Renamed `Aesthetic#processStyleSheet` to `parseStyleSheet`.
 
 #### 🚀 Updates